+++ src/RcsbFvWeb/RcsbFvBuilder/RcsbFvInstanceBuilder.ts
@@ -1,4 +1,4 @@
-import {RcsbFvAdditionalConfig} from "../RcsbFvModule/RcsbFvModuleInterface";
+import {RcsbContextType, RcsbFvAdditionalConfig} from "../RcsbFvModule/RcsbFvModuleInterface";
 import {PolymerEntityInstanceTranslate} from "../../RcsbUtils/PolymerEntityInstanceTranslate";
 import {RcsbFvInstance} from "../RcsbFvModule/RcsbFvInstance";
 import {TagDelimiter} from "../../RcsbUtils/TagDelimiter";
@@ -20,7 +20,7 @@ type instanceModule = "interface"|"instance";
 export interface InstanceSequenceConfig {
     dropdownTitle?: string;
     defaultValue?: string|undefined|null;
-    beforeRenderCallback?:(x: InstanceSequenceOnchangeInterface)=>void;
+    beforeChangeCallback?:(x: InstanceSequenceOnchangeInterface)=>undefined|RcsbContextType;
     onChangeCallback?:(x: InstanceSequenceOnchangeInterface)=>void;
     filterInstances?: Set<string>;
     displayAuthId?: boolean;
@@ -69,17 +69,22 @@ export class RcsbFvInstanceBuilder {
                 shortLabel: config.displayAuthId === true ? instance.authId : label,
                 optId: instance.authId,
                 onChange: async () => {
-                    if (typeof config.beforeRenderCallback === "function")
-                        config.beforeRenderCallback({
+                    let externalContext: RcsbContextType | undefined;
+                    if (typeof config.beforeChangeCallback === "function")
+                        externalContext = config.beforeChangeCallback({
                             pdbId: instance.entryId,
                             authId: instance.authId,
                             asymId: instance.asymId
                         });
+                    const rcsbContext:RcsbContextType = {
+                        ...additionalConfig.rcsbContext,
+                        ...externalContext,
+                        ...instance
+                    };
                     await RcsbFvInstanceBuilder.buildInstanceFv(
                         elementFvId,
                         instance.rcsbId,
-                        additionalConfig,
-                        instance,
+                        {...additionalConfig, rcsbContext: rcsbContext},
                         config.module
                     );
                     if (typeof config.onChangeCallback === "function")
@@ -105,13 +110,19 @@ export class RcsbFvInstanceBuilder {
             label: group[0].groupLabel,
             options: group
         })), {addTitle:true, defaultValue: config.defaultValue, dropdownTitle: (config.dropdownTitle ?? "INSTANCE"), width: config.displayAuthId === true ? 70 : undefined, optionProps: config.selectButtonOptionProps });
-        if (typeof config.beforeRenderCallback === "function")
-            config.beforeRenderCallback({
+        let externalContext: RcsbContextType | undefined;
+        if (typeof config.beforeChangeCallback === "function")
+            externalContext = config.beforeChangeCallback({
                 pdbId: filteredInstanceList[index].entryId,
                 authId: filteredInstanceList[index].authId,
                 asymId: filteredInstanceList[index].asymId
             });
-        const out: RcsbFvModulePublicInterface = await RcsbFvInstanceBuilder.buildInstanceFv(elementFvId, filteredInstanceList[index].rcsbId, additionalConfig, filteredInstanceList[index], config.module);
+        const rcsbContext:RcsbContextType = {
+            ...additionalConfig.rcsbContext,
+            ...externalContext,
+            ...filteredInstanceList[index]
+        };
+        const out: RcsbFvModulePublicInterface = await RcsbFvInstanceBuilder.buildInstanceFv(elementFvId, filteredInstanceList[index].rcsbId, {...additionalConfig, rcsbContext: rcsbContext}, config.module);
         if (typeof config.onChangeCallback === "function")
             config.onChangeCallback({
                 pdbId: filteredInstanceList[index].entryId,
@@ -121,7 +132,8 @@ export class RcsbFvInstanceBuilder {
         return out;
     }
 
-    static async buildInstanceFv(elementId:string, instanceId:string, additionalConfig?:RcsbFvAdditionalConfig, rcsbContext?:PolymerEntityInstanceInterface, module?:instanceModule): Promise<RcsbFvModulePublicInterface> {
+    static async buildInstanceFv(elementId:string, instanceId:string, additionalConfig?:RcsbFvAdditionalConfig, module?:instanceModule): Promise<RcsbFvModulePublicInterface> {
+        console.log(additionalConfig);
         return new Promise<RcsbFvModulePublicInterface>((resolve,reject)=>{
             try {
                 const entryId: string = instanceId.split(TagDelimiter.instance)[0];
@@ -131,7 +143,7 @@ export class RcsbFvInstanceBuilder {
                     additionalConfig: {
                         ...additionalConfig,
                         rcsbContext: {
-                            authId: rcsbContext.authId,
+                            ...additionalConfig?.rcsbContext,
                             entryId: entryId,
                             asymId: asymId
                         }