+++ src/RcsbFvWeb/RcsbFvModule/RcsbFvInterface.ts
@@ -7,7 +7,7 @@ import {
     Type
 } from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes";
 import {RcsbFvAbstractModule} from "./RcsbFvAbstractModule";
-import {RcsbFvModuleBuildInterface} from "./RcsbFvModuleInterface";
+import {RcsbContextType, RcsbFvModuleBuildInterface} from "./RcsbFvModuleInterface";
 import {AnnotationCollectorInterface} from "../../RcsbCollectTools/AnnotationCollector/AnnotationCollectorInterface";
 import {AnnotationCollector} from "../../RcsbCollectTools/AnnotationCollector/AnnotationCollector";
 
@@ -38,7 +38,7 @@ export class RcsbFvInterface extends RcsbFvAbstractModule {
         this.annotationTracks = await this.annotationCollector.collect({
             queryId: instanceId,
             reference: SequenceReference.PdbInstance,
-            titleSuffix: this.titleSuffix.bind(this),
+            titleSuffix: this.titleSuffix(buildConfig?.additionalConfig?.rcsbContext).bind(this),
             trackTitle: this.trackTitle.bind(this),
             typeSuffix: this.typeSuffix.bind(this),
             sources:source,
@@ -66,15 +66,23 @@ export class RcsbFvInterface extends RcsbFvAbstractModule {
         }
     }
 
-    private async titleSuffix(ann: AnnotationFeatures, d: Feature): Promise<string> {
-        if(ann.source === Source.PdbInterface) {
-            const interfaceTranslate = await rcsbFvCtxManager.getInterfaceToInstance(ann.target_id);
-            const chain: string = this.instanceId.split(TagDelimiter.instance)[1];
-            const chPair: [string,string] = interfaceTranslate.getInstances(ann.target_id);
-            const asym: string = chPair[0] == chain ? chPair[1] : chPair[0]
-            const auth:string = (await rcsbFvCtxManager.getEntityToInstance(this.instanceId.split(TagDelimiter.instance)[0])).translateAsymToAuth(asym);
-            return asym == auth ? asym : `${asym}[auth ${auth}]`;
-        }
+    private titleSuffix(rcsbContext: RcsbContextType): ((ann: AnnotationFeatures, d: Feature)=>Promise<string>) {
+        return (async (ann: AnnotationFeatures, d: Feature) => {
+            if (ann.source === Source.PdbInterface) {
+                const interfaceTranslate = await rcsbFvCtxManager.getInterfaceToInstance(ann.target_id);
+                const chain: string = this.instanceId.split(TagDelimiter.instance)[1];
+                const chPair: [string, string] = interfaceTranslate.getInstances(ann.target_id);
+                const asym: string = chPair[0] == chain ? chPair[1] : chPair[0]
+                const auth: string = (await rcsbFvCtxManager.getEntityToInstance(this.instanceId.split(TagDelimiter.instance)[0])).translateAsymToAuth(asym);
+                const operators: [Array<Array<string>>, Array<Array<string>>] = interfaceTranslate.getOperatorIds(ann.target_id);
+                let partnerOperator: string = "";
+                if(Array.isArray(rcsbContext.operatorIds)){
+                    const opIndex: number = operators[ann.target_identifiers.interface_partner_index].map(o=>o.join("-")).indexOf(rcsbContext.operatorIds.join("-"));
+                    partnerOperator = TagDelimiter.operatorComposition+operators[1-ann.target_identifiers.interface_partner_index][opIndex].join(TagDelimiter.operatorComposition);
+                }
+                return (asym == auth ? asym : `${asym}[auth ${auth}]`)+partnerOperator;
+            }
+        });
     }
 
     private async trackTitle(ann: AnnotationFeatures, d: Feature): Promise<string> {